backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alembic=1.14.0=pypi_0
amqp=5.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
appdirs=1.4.4=pypi_0
async-timeout=5.0.1=pypi_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
billiard=4.2.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
celery=5.4.0=pypi_0
certifi=2024.7.4=pypi_0
chardet=5.2.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chemdataextractor=1.3.0=pypi_0
click=8.1.7=pypi_0
click-didyoumean=0.3.1=pypi_0
click-plugins=1.1.1=pypi_0
click-repl=0.3.0=pypi_0
cssselect=1.2.0=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
dawg=0.8.0=pypi_0
ecdsa=0.19.0=pypi_0
exceptiongroup=1.2.2=pypi_0
fastapi=0.112.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.15.4=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.6.1=pypi_0
gensim=4.1.2=pypi_0
gevent=24.2.1=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
huggingface-hub=0.24.5=pypi_0
idna=3.7=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_2
kombu=5.4.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.48.0=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lxml=5.2.2=pypi_0
lz4-c=1.9.4=h6a678d5_1
mako=1.3.6=pypi_0
markupsafe=2.1.5=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nltk=3.8.1=pypi_0
numpy=1.20.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
packaging=24.1=pypi_0
pandas=1.3.4=pypi_0
passlib=1.7.4=pypi_0
pdfdataextractor=1.0=pypi_0
pdfminer-six=20200121=pypi_0
pillow=10.4.0=py38h5eee18b_0
pip=24.0=py38h06a4308_0
prettytable=2.4.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
psycopg2-binary=2.9.10=pypi_0
pyasn1=0.6.0=pypi_0
pycorenlp=0.3.0=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pymupdf=1.24.9=pypi_0
pymupdfb=1.24.9=pypi_0
python=3.8.19=h955ad1f_0
python-crfsuite=0.9.10=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-jose=3.3.0=pypi_0
python-multipart=0.0.9=pypi_0
pytorch=1.10.0=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
redis=5.2.0=pypi_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
rsa=4.9=pypi_0
sacremoses=0.1.1=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.10.1=pypi_0
setuptools=69.5.1=py38h06a4308_0
six=1.16.0=pypi_0
smart-open=7.0.4=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.32=pypi_0
sqlite=3.45.3=h5eee18b_0
starlette=0.37.2=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.10.3=pypi_0
torchaudio=0.10.0=py38_cu113
torchvision=0.11.1=py38_cu113
tqdm=4.66.5=pypi_0
transformers=4.13.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.2=pypi_0
ujson=5.10.0=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.5=pypi_0
vine=5.1.0=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py38h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
zope-event=5.0=pypi_0
zope-interface=7.1.1=pypi_0
zstd=1.5.5=hc292b87_2
